web-app/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
torch>=2.0.0
diffusers @ git+https://github.com/huggingface/diffusers
transformers
accelerate
pillow
python-multipart
